Fix entities having integer positions

+++ source/entity.hpp
@@ -7,12 +7,12 @@ enum EntityType { ENT_DUMMY = 0, ENT_PLAYER, ENT_AI };
 class Entity
 {
 public:
-	int x, y;
+	double x, y;
 	int health;
 	EntityType type;
 
 	Entity() : x(0), y(0), health(1), type(ENT_DUMMY) {};
-	virtual void Update(double dt, double time) { x += 100*dt; };
+	virtual void Update(double dt, double time) { x += 100.0*dt; };
 };
 
 #endif
